Remove subscription contracts (#397)
* rm: sub contracts

* rm: sub offchain

fix: abi & test references

* Configure toll onramp

* fix commitstore config

* Set perm seconds

+++ contracts/test/v0.8/foundry/dev/special/MockLinkToken.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 contract MockLinkToken {
-  uint public constant totalSupply = 10**27;
+  uint256 public constant totalSupply = 10**27;
 
   mapping(address => uint256) public balances;
 
@@ -11,17 +11,21 @@ contract MockLinkToken {
   }
 
   /**
-  * @dev transfer token for a specified address
-  * @param _to The address to transfer to.
-  * @param _value The amount to be transferred.
-  */
+   * @dev transfer token for a specified address
+   * @param _to The address to transfer to.
+   * @param _value The amount to be transferred.
+   */
   function transfer(address _to, uint256 _value) public returns (bool) {
     balances[msg.sender] = balances[msg.sender] - _value;
     balances[_to] = balances[_to] + _value;
     return true;
   }
 
-  function transferAndCall(address _to, uint _value, bytes calldata _data) public validRecipient(_to) returns (bool success) {
+  function transferAndCall(
+    address _to,
+    uint256 _value,
+    bytes calldata _data
+  ) public validRecipient(_to) returns (bool success) {
     transfer(_to, _value);
     if (isContract(_to)) {
       contractFallback(_to, _value, _data);
@@ -38,23 +42,28 @@ contract MockLinkToken {
     _;
   }
 
-  function contractFallback(address _to, uint _value, bytes calldata _data)
-    private
-  {
+  function contractFallback(
+    address _to,
+    uint256 _value,
+    bytes calldata _data
+  ) private {
     ERC677Receiver receiver = ERC677Receiver(_to);
     receiver.onTokenTransfer(msg.sender, _value, _data);
   }
 
-  function isContract(address _addr)
-    private
-    returns (bool hasCode)
-  {
-    uint length;
-    assembly { length := extcodesize(_addr) }
+  function isContract(address _addr) private returns (bool hasCode) {
+    uint256 length;
+    assembly {
+      length := extcodesize(_addr)
+    }
     return length > 0;
   }
 }
 
 interface ERC677Receiver {
-  function onTokenTransfer(address _sender, uint _value, bytes calldata _data) external;
+  function onTokenTransfer(
+    address _sender,
+    uint256 _value,
+    bytes calldata _data
+  ) external;
 }
